+++ crates/data_node/src/api/accounts.rs
@@ -1,7 +1,7 @@
 use crate::client::DataNode;
 use bf_api_provider::types::{
     AccountsAddressesResponse, AccountsDelegationsResponse, AccountsRegistrationsResponse,
-    AccountsResponse, AccountsRewardsResponse,
+    AccountsResponse, AccountsRewardsResponse, AccountsWithdrawalsResponse,
 };
 use bf_common::{pagination::Pagination, types::ApiResult};
 
@@ -58,6 +58,16 @@ impl DataNodeAccounts<'_> {
         pagination: &Pagination,
     ) -> ApiResult<AccountsRegistrationsResponse> {
         let path = format!("accounts/{stake_address}/registrations");
+
+        self.inner.client.get(&path, Some(pagination)).await
+    }
+
+    pub async fn withdrawals(
+        &self,
+        stake_address: &str,
+        pagination: &Pagination,
+    ) -> ApiResult<AccountsWithdrawalsResponse> {
+        let path = format!("accounts/{stake_address}/withdrawals");
 
         self.inner.client.get(&path, Some(pagination)).await
     }