+++ crates/platform/src/api/pools/pool_id/root.rs
@@ -1,6 +1,17 @@
-use crate::{BlockfrostError, api::ApiResult};
+use crate::{
+    api::ApiResult,
+    pools::{PoolData, PoolsPath},
+    server::state::AppState,
+};
+use axum::extract::{Path, State};
 use bf_api_provider::types::PoolsSingleResponse;
 
-pub async fn route() -> ApiResult<PoolsSingleResponse> {
-    Err(BlockfrostError::not_found())
+pub async fn route(
+    State(state): State<AppState>,
+    Path(pools_path): Path<PoolsPath>,
+) -> ApiResult<PoolsSingleResponse> {
+    let pool_data = PoolData::from_path(&pools_path.pool_id)?;
+    let data_node = state.data_node()?;
+
+    data_node.pools().by_id(&pool_data.pool_id).await
 }